Additional lines to inject into the generated dnsmasq configuration.
Additional lines to inject into the generated dnsmasq configuration.
Warning: This is an advanced setting and should only be used with care. Incorrectly formatted or duplicated lines as well as lines conflicting with the automatic configuration of Pi-hole can break the embedded dnsmasq and will stop DNS resolution from working.