Static DHCP configuration - Specify per host parameters for the DHCP server.
Static DHCP configuration    

Specify per host parameters for the DHCP server. This allows a machine with a particular hardware address to be always allocated the same hostname, IP address and lease time. A hostname specified like this overrides any supplied by the DHCP client on the machine. It is also allowable to omit the hardware address and include the hostname, in which case the IP address and lease times will apply to any machine claiming that name.

 

Each entry should be on a separate line, and should be of the form:

[<hwaddr>][,id:<client_id>|*][,set:<tag>][,tag:<tag>][,<ipaddr>][,<hostname>][,<lease_time>][,ignore]

Only one entry per MAC address is allowed.

Examples:

    00:20:e0:3b:13:af,192.168.0.123

    tells Pi-hole to give the machine with hardware address 00:20:e0:3b:13:af the address 192.168.0.123
     

    00:20:e0:3b:13:af,laptop

    tells Pi-hole to give the machine with hardware address 00:20:e0:3b:13:af the name laptop
     

    00:20:e0:3b:13:af,192.168.0.123,laptop,infinite

    tells Pi-hole to give the machine with hardware address 00:20:e0:3b:13:af the address 192.168.0.123, the name laptop, and an infinite DHCP lease